+++ frontend/src/config/constants.ts
@@ -4,3 +4,8 @@ export const WORDLE_TOKEN_ADDRESS = import.meta.env.VITE_WORDLE_TOKEN_ADDRESS;
 export const WORDLE_GAME_ADDRESS = import.meta.env.VITE_WORDLE_GAME_ADDRESS;
 export const TOKEN_DECIMALS = 18;
 export const NUMBER_OF_GUESSES = 5;
+export const LETTER_BG_COLORS = {
+  2: "#538D4E",
+  1: "#B59F3B",
+  0: "#3A3A3C"
+};
